refactor: move sideSection into components folder
