+++ search-app/src/components/main/sideSection/styles.css
